+++ src/components/content/allergies/helpers/history.tsx
@@ -1,3 +1,4 @@
+import { AllergyIntolerance } from "fhir/r4";
 import { SearchParams } from "fhir-kit-client";
 import { HistoryEntryProps } from "../../resource/helpers/history-entry";
 import { useHistory } from "../../resource/history";
@@ -5,7 +6,7 @@ import { AllergyModel } from "@/fhir/models/allergies";
 import { capitalize } from "@/utils/nodash";
 import { QUERY_KEY_ALLERGY_HISTORY } from "@/utils/query-keys";
 
-export function useAllergiesHistory(enableFQS: boolean, allergy: AllergyModel) {
+export function useAllergiesHistory(allergy: AllergyModel) {
   return useHistory({
     resourceType: "AllergyIntolerance",
     model: allergy,
@@ -14,8 +15,7 @@ export function useAllergiesHistory(enableFQS: boolean, allergy: AllergyModel) {
     valuesToDedupeOn,
     getSearchParams,
     getHistoryEntry,
-    getFiltersFQS,
-    enableFQS,
+    clientSideFiltersFQS,
   });
 }
 
@@ -37,16 +37,15 @@ function getSearchParams(allergy: AllergyModel) {
   return searchParams;
 }
 
-function getFiltersFQS(allergy: AllergyModel) {
-  const tokens = allergy.knownCodings.map((coding) => `${coding.system}|${coding.code}`);
-
-  if (tokens.length > 0) {
-    return {
-      code: { anymatch: tokens },
-    };
-  }
-
-  return undefined;
+function clientSideFiltersFQS(model: AllergyModel, allergies: AllergyIntolerance[]) {
+  const tokens = model.knownCodings.map((coding) => `${coding.system}|${coding.code}`);
+  return allergies.filter((allergy) => {
+    const modelTokens = allergy.code?.coding?.map((coding) => `${coding.system}|${coding.code}`);
+    // Sometimes data doesn't have known codings, so we also match on the code.text making sure that it is not blank.
+    const matchingCodeDisplay = model.codeText && allergy.code?.text === model.codeText;
+    const matchingSystemCode = modelTokens?.some((token) => tokens.includes(token));
+    return matchingSystemCode || matchingCodeDisplay;
+  });
 }
 
 function getHistoryEntry(allergy: AllergyModel): HistoryEntryProps {