+++ src/components/content/medications/patient-medications-outside-badge.tsx
@@ -1,20 +1,14 @@
-import { Badge } from "@/components/core/badge";
+import { UnreadNotification } from "@/components/core/unread-notification";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
 
 export const PatientMedicationsOutsideBadge = () => {
   const { otherProviderMedications = [] } = useQueryAllPatientMedications();
   const activeUnarchivedMedications = otherProviderMedications.filter(
-    (medication) => !(medication.isArchived || medication.isInactive)
+    (medication) => !(medication.isDismissed || medication.isInactive)
   );
 
   if (activeUnarchivedMedications.length > 0) {
-    return (
-      <Badge
-        color="notification"
-        text={activeUnarchivedMedications.length.toString()}
-        className="ctw-h-5"
-      />
-    );
+    return <UnreadNotification />;
   }
   return null;
 };