+++ src/components/content/zus-aggregated-profile/zus-aggregated-profile-tabs.tsx
@@ -1,8 +1,10 @@
+import { UnreadAllergiesNotification } from "../allergies/unread-allergies-notification";
 import {
   PatientConditionsOutside,
   PatientConditionsOutsideProps,
 } from "../conditions/patient-conditions-outside";
 import { PatientConditionsOutsideBadge } from "../conditions/patient-conditions-outside-badge";
+import { UnreadImmunizationsNotification } from "../immunizations/unread-immunizations-notification";
 import { PatientMedications, PatientMedicationsProps } from "../medications/patient-medications";
 import {
   PatientMedicationsOutside,
@@ -32,8 +34,8 @@ import {
   PatientConditionsProps,
 } from "@/components/content/conditions/patient-conditions";
 import {
-  PatientDocumentProps,
   PatientDocuments,
+  PatientDocumentsProps,
 } from "@/components/content/document/patient-documents";
 import {
   PatientImmunizations,
@@ -50,7 +52,12 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   allergies: (props: PatientAllergiesProps = {}) => ({
     key: "allergies",
     getPanelClassName: () => "ctw-pt-5",
-    display: () => "allergies",
+    display: () => (
+      <div className="ctw-flex ctw-items-center ctw-space-x-2">
+        <UnreadAllergiesNotification />
+        <span className="ctw-capitalize">allergies</span>
+      </div>
+    ),
     render: () => <PatientAllergies {...props} />,
   }),
 
@@ -74,15 +81,15 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   "conditions-outside": (props: PatientConditionsOutsideProps = {}) => ({
     key: "conditions-outside",
     display: () => (
-      <div className="ctw-space-x-2">
-        <span className="ctw-capitalize">{i18next.t("zap.tabs.conditionsOutside")}</span>
+      <div className="ctw-flex ctw-items-center ctw-space-x-2">
         <PatientConditionsOutsideBadge />
+        <span className="ctw-capitalize">{i18next.t("zap.tabs.conditionsOutside")}</span>
       </div>
     ),
     render: () => <PatientConditionsOutside {...props} />,
   }),
 
-  documents: (props: PatientDocumentProps = {}) => ({
+  documents: (props: PatientDocumentsProps = {}) => ({
     key: "documents",
     getPanelClassName: () => "ctw-pt-5",
     display: () => (
@@ -98,7 +105,8 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
     key: "immunizations",
     getPanelClassName: () => "ctw-pt-5",
     display: () => (
-      <div className="ctw-space-x-1">
+      <div className="ctw-flex ctw-items-center ctw-space-x-2">
+        <UnreadImmunizationsNotification />
         <span className="ctw-capitalize">immunizations</span>
         <BetaLabel />
       </div>
@@ -115,9 +123,9 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   "medications-outside": (props: PatientMedicationsOutsideProps = {}) => ({
     key: "medications-outside",
     display: () => (
-      <div className="ctw-space-x-2">
-        <span className="ctw-capitalize">{i18next.t("zap.tabs.medicationsOutside")}</span>
+      <div className="ctw-flex ctw-items-center ctw-space-x-2">
         <PatientMedicationsOutsideBadge />
+        <span className="ctw-capitalize">{i18next.t("zap.tabs.medicationsOutside")}</span>
       </div>
     ),
     render: () => <PatientMedicationsOutside {...props} />,