+++ src/components/core/providers/ctw-context.tsx
@@ -20,6 +20,7 @@ export type CTWRequestContext = {
   authToken: string;
   // The user's builder ID.
   builderId: string;
+  userType: string | undefined;
   // The optional builder ID used in case the user is impersonating a builder.
   contextBuilderId: string | undefined;
   fhirClient: Client;