+++ src/fhir/models/fhir-model.ts
@@ -1,6 +1,13 @@
 import { Basic, Resource } from "fhir/r4";
-import { SYSTEM_ENRICHMENT, SYSTEM_SUMMARY, SYSTEM_ZUS_PROFILE_ACTION } from "../system-urls";
-import { ResourceMap, ResourceTypeString } from "../types";
+import {
+  SYSTEM_ENRICHMENT,
+  SYSTEM_SUMMARY,
+  SYSTEM_ZUS_OWNER,
+  SYSTEM_ZUS_PROFILE_ACTION,
+  SYSTEM_ZUS_THIRD_PARTY,
+  SYSTEM_ZUS_UNIVERSAL_ID,
+} from "../system-urls";
+import { isFHIRDomainResource, ResourceMap, ResourceTypeString } from "../types";
 import { find, orderBy, some, startCase } from "@/utils/nodash";
 
 export abstract class FHIRModel<T extends fhir4.Resource> {
@@ -33,7 +40,7 @@ export abstract class FHIRModel<T extends fhir4.Resource> {
     return this.resource.meta?.versionId || "";
   }
 
-  get isArchived(): boolean {
+  get isDismissed(): boolean {
     const basic = this.getLatestBasicResourceByActions(["archive", "unarchive"]);
     return some(basic?.code.coding, {
       system: SYSTEM_ZUS_PROFILE_ACTION,
@@ -41,9 +48,24 @@ export abstract class FHIRModel<T extends fhir4.Resource> {
     });
   }
 
+  get isRead(): boolean {
+    const basic = this.getLatestBasicResourceByActions(["read", "unread"]);
+    return some(basic?.code.coding, {
+      system: SYSTEM_ZUS_PROFILE_ACTION,
+      code: "read",
+    });
+  }
+
   // Returns true if this resource is a summary/lens resource.
   get isSummaryResource(): boolean {
     return find(this.resource.meta?.tag, { system: SYSTEM_SUMMARY }) !== undefined;
+  }
+
+  get patientUPID(): string | undefined {
+    if (isFHIRDomainResource(this.resource)) {
+      return find(this.resource.extension, { url: SYSTEM_ZUS_UNIVERSAL_ID })?.valueString;
+    }
+    return undefined;
   }
 
   get resourceType(): ResourceTypeString {
@@ -86,6 +108,20 @@ export abstract class FHIRModel<T extends fhir4.Resource> {
   // from Zus enrichment.
   isEnriched(): boolean {
     return JSON.stringify(this.resource).includes(SYSTEM_ENRICHMENT);
+  }
+
+  // Returns true if this resource is owned by the specified builder.
+  ownedByBuilder(builderId: string): boolean {
+    if (this.resource.meta?.tag?.some((tag) => tag.system === SYSTEM_ZUS_THIRD_PARTY)) {
+      return false;
+    }
+    const ownerTag = this.resource.meta?.tag?.find((tag) => tag.system === SYSTEM_ZUS_OWNER);
+    const splitTag = ownerTag?.code?.split("/");
+
+    if (splitTag?.length !== 2) {
+      return false;
+    }
+    return splitTag[1] === builderId;
   }
 
   // Returns a string that would setup this model.