+++ src/fhir/provenance.ts
@@ -1,9 +1,16 @@
 import { Provenance, Resource } from "fhir/r4";
+import { getResources } from "./bundle";
 import { FHIRModel } from "./models/fhir-model";
 import { getUsersPractitionerReference } from "./practitioner";
 import { searchAllRecords } from "./search-helpers";
-import { SYSTEM_PROVENANCE_ACTIVITY_TYPE, SYSTEM_PROVENANCE_AGENT_TYPE } from "./system-urls";
+import {
+  SYSTEM_PROVENANCE_ACTIVITY_TYPE,
+  SYSTEM_PROVENANCE_AGENT_TYPE,
+  SYSTEM_ZUS_UNIVERSAL_ID,
+} from "./system-urls";
+import { ResourceTypeString } from "./types";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
+import { searchProvenancesFQS } from "@/services/fqs/queries/provenances";
 import { claimsBuilderName } from "@/utils/auth";
 import { uniq } from "@/utils/nodash";
 import { QUERY_KEY_PROVENANCE } from "@/utils/query-keys";
@@ -84,26 +91,34 @@ export const createProvenance = async (
 export async function searchProvenances<T extends fhir4.Resource>(
   requestContext: CTWRequestContext,
   models: FHIRModel<T>[],
-  // TODO remove underscore prefix when used.
-  _enableFQS = false
+  enableFQS = false
 ): Promise<Provenance[]> {
   if (models.length === 0) return [];
 
   const targets = uniq(models.map((m) => `${m.resourceType}/${m.id}`));
+  const ids = models.map((m) => m.id);
 
-  // TODO uncomment when FQS supports provenance on all resources and is back-filled.
-  // if (enableFQS) {
-  //   return searchProvenancesFQS(requestContext, models[0].resourceType, targets);
-  // }
-  return searchProvenancesODS(requestContext, targets);
+  if (enableFQS) {
+    return searchProvenancesFQS(requestContext, models[0].resourceType, targets);
+  }
+  return searchProvenancesODS(requestContext, ids, models[0].resourceType, models[0].patientUPID);
 }
 
-export async function searchProvenancesODS(requestContext: CTWRequestContext, targets: string[]) {
-  const target = targets.join(",");
-  const { resources } = await queryClient.fetchQuery([QUERY_KEY_PROVENANCE, target], async () =>
-    searchAllRecords("Provenance", requestContext, {
-      target,
+export async function searchProvenancesODS(
+  requestContext: CTWRequestContext,
+  ids: string[],
+  resourceType: ResourceTypeString,
+  patientUPID: string | undefined
+) {
+  const id = ids.join(",");
+  const { bundle } = await queryClient.fetchQuery([QUERY_KEY_PROVENANCE, id], async () =>
+    searchAllRecords(resourceType, requestContext, {
+      _id: id,
+      _revinclude: "Provenance:target",
+      // Need to include patient identifier to work around CPR.
+      "patient.identifier": patientUPID ? `${SYSTEM_ZUS_UNIVERSAL_ID}|${patientUPID}` : "",
     })
   );
-  return resources;
+
+  return getResources(bundle, "Provenance");
 }