+++ src/services/fqs/queries/medication-request.ts
@@ -36,6 +36,10 @@ export const medicationRequestQuery = gql`
         node {
           id
           resourceType
+          extension {
+            url
+            valueString
+          }
           authoredOn
           contained {
             resource {
@@ -102,7 +106,35 @@ export const medicationRequestQuery = gql`
             ...Reference
           }
           requester {
-            ...Reference
+            id
+            extension {
+              url
+              valueString
+            }
+            reference
+            type
+            identifier {
+              id
+              system
+              value
+            }
+            display
+            resource {
+              ... on Practitioner {
+                id
+                resourceType
+                name {
+                  family
+                  given
+                }
+                address {
+                  text
+                }
+                telecom {
+                  value
+                }
+              }
+            }
           }
           status
         }